fixed connection to mongo
