+++ cradlepoint/package.json
@@ -24,7 +24,8 @@
     "react-modal": "^3.14.3"
   },
   "devDependencies": {
-    "eslint": "7.32.0",
+    "@babel/types": "^7.16.0",
+    "eslint": "^7.32.0",
     "eslint-config-next": "11.1.2"
   }
 }
